requirements.txt
python-telegram-bot==21.4
requests==2.31.0
openai==0.28